Update Hook UserBot to version 0.8
A fairly global update for the Hook UserBot. We've redesigned the basic mechanics of the project, fixed the installer, added new features, and optimized the code. Also, various errors of previous versions have been fixed.

We remind you that all versions <1.0 are testing versions, that is, they may contain errors. If you don't agree to test the product or you don't like the bugs, wait for version 1.0 (the first stable version).