Another update has been carried out, which makes our path closer to the goal - version 1.0.0!
What has been changed: fixed bugs from the previous version, updated and improved the interface (including the web interface), fixed an error during installation, and we are also fixing translations into other languages (currently only Russian and English languages).  We remind you that this is still a beta version, and therefore it may (and most likely) contain errors. If you want to see a stable and finished product, please expect the first stable version (1.0.0). If you find an error or bug, please report it in the "Issues" section. We will be grateful.