Pre-release v. 1.0.0.
Well. Our path is slowly entering a new path - version 0.9.0! This version is a direct candidate for the release of version 1.0.0! Here, we have improved and tested all the necessary functions, urgently fixed all the errors found, and also fixed the critical vulnerabilities of the kernel that we managed to find.