Release of version 1.0.0 has arrived
Well. So the test release of version 1.0.0 has arrived, which our entire development team has been working tirelessly on from morning until late at night. We are pleased to inform you that the current version is stable (or those errors that exist are uncritical and we are monitoring them). A little later, we will post a beautiful and accessible, understandable README.md which will become the installation instructions, and later we will develop the documentation!)